# ao/co IndexOnlyScan: Use the last cached minipage