We can use the last cached minipage in the AppendOnlyBlockDirectory
structure during successive index only checks.

Today, index only scans on AO/CO tables have enormous per-tuple overhead
borne from the cost of starting, executing and tearing down short-lived
block directory sysscans.

This patch can result in a dramatic reduction in the number of these
shortlived sysscans, by referring to the last cached minipage. The best
case would be for highly correlated data where the next lookup would be
satisfied from the last cached minipage, with very high probability.

Example (TPC-H AO table of on-disk size 26GB, scale=30):

explain analyze select count(*) from lineitem where l_shipdate >= date '1994-01-01'
    and l_shipdate < date '1994-01-01' + interval '1' year;

and we have a clustered b-tree index on l_shipdate

Without patch: statement_timeout of 5min.
With patch: the query takes 3s to execute.

Dev-pipeline: https://dev.ci.gpdb.pivotal.io/teams/main/pipelines/optimize_blkdir_ios?group=all