@soumyadeep2007 2023-09-09
without patch:

5min statement_timeout.

with patch:

brin=# explain analyze select count(*) from lineitem where l_shipdate >= date '1994-01-01'
                                        and l_shipdate < date '1994-01-01' + interval '1' year;
                                                                                 QUERY PLAN                                                                                  
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=444955.63..444955.64 rows=1 width=8) (actual time=3160.754..3160.755 rows=1 loops=1)
   ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=444955.57..444955.62 rows=3 width=8) (actual time=3121.888..3160.744 rows=3 loops=1)
         ->  Partial Aggregate  (cost=444955.57..444955.58 rows=1 width=8) (actual time=3154.715..3154.715 rows=1 loops=1)
               ->  Index Only Scan using lineitem_l_shipdate_idx on lineitem  (cost=0.19..422642.43 rows=8925256 width=0) (actual time=9.301..2581.600 rows=9103540 loops=1)
                     Index Cond: ((l_shipdate >= '1994-01-01'::date) AND (l_shipdate < '1995-01-01 00:00:00'::timestamp without time zone))
                     Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 3.017 ms
   (slice0)    Executor memory: 43K bytes.
   (slice1)    Executor memory: 179K bytes avg x 3 workers, 179K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 3178.048 ms
(12 rows)