@soumyadeep2007 2023-09-10
More data points:

(1) For the original non-clustered case: https://github.com/greenplum-db/gpdb/issues/15914#issue-1787326778

Without patch:

```
haolin=# explain analyze select count(*) from t2i where a > 991243 and a < 1045762;
                                                                        QUERY PLAN                                                                         
-----------------------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=820363.33..820363.34 rows=1 width=8) (actual time=97524.519..97524.520 rows=1 loops=1)
   ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=820363.27..820363.32 rows=3 width=8) (actual time=96475.838..97524.507 rows=3 loops=1)
         ->  Partial Aggregate  (cost=820363.27..820363.28 rows=1 width=8) (actual time=97164.528..97164.529 rows=1 loops=1)
               ->  Index Only Scan using t2i_a_idx on t2i  (cost=0.20..814830.80 rows=2212989 width=0) (actual time=5.688..97150.656 rows=2335104 loops=1)
                     Index Cond: ((a > 991243) AND (a < 1045762))
                     Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 15.193 ms
   (slice0)    Executor memory: 43K bytes.
   (slice1)    Executor memory: 179K bytes avg x 3 workers, 179K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 97555.784 ms
(12 rows)
```


With patch:

```
haolin=# explain analyze select count(*) from t2i where a > 991243 and a < 1045762;
                                                                        QUERY PLAN                                                                         
-----------------------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=112112.29..112112.30 rows=1 width=8) (actual time=98991.020..98991.021 rows=1 loops=1)
   ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=112112.23..112112.28 rows=3 width=8) (actual time=97839.873..98991.007 rows=3 loops=1)
         ->  Partial Aggregate  (cost=112112.23..112112.24 rows=1 width=8) (actual time=98986.559..98986.560 rows=1 loops=1)
               ->  Index Only Scan using t2i_a_idx on t2i  (cost=0.20..106579.76 rows=2212989 width=0) (actual time=5.385..98273.373 rows=2335104 loops=1)
                     Index Cond: ((a > 991243) AND (a < 1045762))
                     Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 4.035 ms
   (slice0)    Executor memory: 43K bytes.
   (slice1)    Executor memory: 179K bytes avg x 3 workers, 179K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 99007.524 ms
(12 rows)
```

(2) For the clustered case here: https://github.com/greenplum-db/gpdb/issues/15914#issuecomment-1711659596

Without patch:
```
haolin=# explain analyze select * from ao1 where a > 991243 and a < 1045762;
                                                             QUERY PLAN                                                              
-------------------------------------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.19..1623.97 rows=55192 width=4) (actual time=6.143..91.660 rows=54518 loops=1)
   ->  Index Only Scan using ao1_a_idx on ao1  (cost=0.19..888.08 rows=18397 width=4) (actual time=2.579..85.267 rows=18243 loops=1)
         Index Cond: ((a > 991243) AND (a < 1045762))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 2.019 ms
   (slice0)    Executor memory: 39K bytes.
   (slice1)    Executor memory: 171K bytes avg x 3 workers, 171K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 104.179 ms
(10 rows)

```

With patch:

```
haolin=# explain analyze select * from ao1 where a > 991243 and a < 1045762;
                                                             QUERY PLAN                                                             
------------------------------------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.19..1623.97 rows=55192 width=4) (actual time=1.033..12.346 rows=54518 loops=1)
   ->  Index Only Scan using ao1_a_idx on ao1  (cost=0.19..888.08 rows=18397 width=4) (actual time=0.379..5.356 rows=18243 loops=1)
         Index Cond: ((a > 991243) AND (a < 1045762))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 3.832 ms
   (slice0)    Executor memory: 39K bytes.
   (slice1)    Executor memory: 150K bytes avg x 3 workers, 150K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 14.240 ms
(10 rows)

```