@haolinw 2023-09-11
LGTM. Probably for later minor enhancement, we may want to re-use `cached_mpentry_num` to cache the current entry_no and check the rowNum whether falls into the range between minipageInfo->minipage->entry[entry_no].firstRowNum and firstRowNum + rowCount - 1 to reduce some `find_minipage_entry` calls.