@soumyadeep2007 2023-09-11
> Probably for later minor enhancement, we may want to re-use cached_mpentry_num to cache the current entry_no and check the rowNum whether falls into the range between minipageInfo->minipage->entry[entry_no].firstRowNum and firstRowNum + rowCount - 1 to reduce some find_minipage_entry calls.

That is an excellent suggestion! What we can do is change `find_minipage_entry` to have signature:
`find_minipage_entry(AppendOnlyBlockDirectory *blockDirectory, int columnGroupNo, int64 rowNum)`

and then inside we can check if `blockDirectory->cached_mpentry_num != InvalidEntryNum`. If it is invalid, we don't care about it. If it is not invalid, then we check  between `minipageInfo->minipage->entry[cached_mpentry_num].firstRowNum` and `minipageInfo->minipage->entry[cached_mpentry_num].firstRowNum + rowCount - 1`

And at the end of `find_minipage_entry` we can set `blockDirectory->cached_mpentry_num` if we found a valid entry.

This way both `AppendOnlyBlockDirectory_GetEntry` and `blkdir_entry_exists` can benefit.

Let's do this in a follow-up PR as a further enhancement!