@haolinw 2023-09-19
> Let's do this in a follow-up PR as a further enhancement!

Opened a PR https://github.com/greenplum-db/gpdb/pull/16456 to address this.